backend/requirements.txt
# Core dependencies
Flask
requests>=2.20.0
google-cloud-secret-manager>=2.16.0
google-cloud-storage==2.10.0
vertexai
ffmpeg-python
google-cloud-aiplatform
gunicorn

# YouTube API dependencies
google-api-python-client>=2.80.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
python-dotenv>=1.0.0

# Cloud Functions dependencies
functions-framework

# Testing dependencies
pytest==7.4.0
pytest-mock==3.11.1
pytest-cov==4.1.0

# Data validation
pydantic>=2.0.0

# Development tools - versions pinned for consistency
black==24.1.1        # Code formatter
ruff==0.2.1         # Fast linter and import sorter
